fitur 9,13

+++ SI-FACTORY/src/main/java/apap/group/assignment/SIFACTORY/model/DeliveryModel.java
@@ -28,7 +28,7 @@ public class DeliveryModel implements Serializable{
 
     @NotNull
     @Column(nullable = false)
-    private Integer idKurir;
+    private Long idKurir;
 
     @NotNull
     @Column(nullable = false)
